+++ Source/ACE.Database/Models/Shard/BiotaPropertiesPosition.cs
@@ -7,7 +7,7 @@ namespace ACE.Database.Models.Shard
 {
     public partial class BiotaPropertiesPosition
     {
-        public uint ObjectId { get; set; }
+        public ulong ObjectId { get; set; }
         public ushort PositionType { get; set; }
         public uint? Instance { get; set; }
         public uint ObjCellId { get; set; }
